+++ hydra-node/src/Hydra/Chain.hs
@@ -36,6 +36,7 @@ import Hydra.Tx (
   IsTx (..),
   SnapshotNumber,
   SnapshotVersion,
+  TxOutType,
   UTxOType,
  )
 import Hydra.Tx.OnChainId (OnChainId)
@@ -148,7 +149,7 @@ data OnChainTx tx
       , contestationDeadline :: UTCTime
       }
   | OnFanoutTx {headId :: HeadId, fanoutUTxO :: UTxOType tx}
-  | OnPartialFanoutTx {headId :: HeadId, distributedUTxO :: UTxOType tx}
+  | OnPartialFanoutTx {headId :: HeadId, distributedOutputs :: Set (TxOutType tx)}
   deriving stock (Generic)
 
 deriving stock instance IsTx tx => Eq (OnChainTx tx)