+++ hydra-node/src/Hydra/HeadLogic/Outcome.hs
@@ -23,6 +23,7 @@ import Hydra.Tx (
   SnapshotNumber,
   SnapshotVersion,
   TxIdType,
+  TxOutType,
   UTxOType,
  )
 import Hydra.Tx.ContestationPeriod (ContestationPeriod)
@@ -124,11 +125,11 @@ data StateChanged tx
   | HeadClosed {headId :: HeadId, snapshotNumber :: SnapshotNumber, chainState :: ChainStateType tx, contestationDeadline :: UTCTime}
   | HeadContested {headId :: HeadId, chainState :: ChainStateType tx, contestationDeadline :: UTCTime, snapshotNumber :: SnapshotNumber}
   | HeadIsReadyToFanout {headId :: HeadId}
-  | HeadFannedOut {headId :: HeadId, utxo :: UTxOType tx, chainState :: ChainStateType tx}
+  | HeadFannedOut {headId :: HeadId, finalizedOutputs :: Set (TxOutType tx), chainState :: ChainStateType tx}
   | HeadPartialFannedOut
       { headId :: HeadId
-      , distributedUTxO :: UTxOType tx
-      , remainingUTxO :: UTxOType tx
+      , distributedOutputs :: Set (TxOutType tx)
+      , remainingOutputs :: Set (TxOutType tx)
       , chainState :: ChainStateType tx
       }
   | ChainRolledBack {chainState :: ChainStateType tx}