+++ hydra-node/test/Hydra/Chain/Direct/HandlersSpec.hs
@@ -11,7 +11,6 @@ import Hydra.Cardano.Api (
   ChainPoint (..),
   SlotNo (..),
   Tx,
-  UTxO,
   fromLedgerTx,
   getChainPoint,
   toLedgerTx,
@@ -50,7 +49,6 @@ import Hydra.HeadLogic (fanoutChunkSize, fanoutOutputThreshold)
 import Hydra.Ledger.Cardano.Time (slotNoToUTCTime)
 import Hydra.Tx (mkSimpleBlueprintTx)
 import Hydra.Tx.Deposit (depositTx)
-import Hydra.Tx.IsTx (splitUTxOAt)
 import Hydra.Tx.Observe (InitObservation (..), observeInitTx)
 import Test.Hydra.Chain ()
 import Test.Hydra.Chain.Direct.State (
@@ -279,35 +277,6 @@ spec = do
       run $ forM_ blocksAfter $ \(TestBlock header txs) -> onRollForward resumedHandler header txs
       latestResumedChainState <- run . atomically $ getLatest resumedLocalChainState
       pure $ latestResumedChainState === latestChainState
-
-  describe "splitUTxOAt (IsTx)" $ do
-    it "splits UTxO into n and remaining" $ do
-      let utxo = generateWith (arbitrary `suchThat` \u -> UTxO.size u > 3) 42 :: UTxO
-          n = 2
-          (first', rest) = splitUTxOAt n utxo
-      UTxO.size first' `shouldBe` n
-      UTxO.size rest `shouldBe` (UTxO.size utxo - n)
-
-    it "preserves all entries" $ do
-      let utxo = generateWith (arbitrary `suchThat` \u -> UTxO.size u > 3) 42 :: UTxO
-          n = 2
-          (first', rest) = splitUTxOAt n utxo
-      UTxO.toList (first' <> rest) `shouldBe` UTxO.toList utxo
-
-    it "handles n larger than UTxO size" $ do
-      let fullUtxo = generateWith (arbitrary `suchThat` \u -> UTxO.size u > 3) 42 :: UTxO
-          -- Take only the first 3 entries to get a small UTxO
-          utxo = UTxO.fromList $ take 3 (UTxO.toList fullUtxo)
-          n = 100
-          (first', rest) = splitUTxOAt n utxo
-      UTxO.size first' `shouldBe` UTxO.size utxo
-      UTxO.size rest `shouldBe` 0
-
-    it "handles empty UTxO" $ do
-      let utxo = mempty :: UTxO
-          (first', rest) = splitUTxOAt 5 utxo
-      UTxO.size first' `shouldBe` 0
-      UTxO.size rest `shouldBe` 0
 
   describe "fanout constants" $ do
     it "chunk size is less than threshold" $