+++ hydra-node/test/Hydra/Model.hs
@@ -150,7 +150,7 @@ instance StateModel WorldState where
     Close :: {party :: Party} -> Action WorldState ()
     -- NOTE: No records possible here as we would duplicate 'Party' fields with
     -- different return values.
-    Fanout :: Party -> Action WorldState UTxO
+    Fanout :: Party -> Action WorldState (Set (TxOut CtxUTxO))
     NewTx :: Party -> Payment -> Action WorldState Payment
     Wait :: DiffTime -> Action WorldState ()
     ObserveConfirmedTx :: Var Payment -> Action WorldState ()
@@ -495,7 +495,7 @@ instance
     case action of
       Fanout{} ->
         case hydraState st of
-          Final{finalUTxO} -> sortTxOuts (toTxOuts finalUTxO) === sortTxOuts (UTxO.txOutputs result)
+          Final{finalUTxO} -> sortTxOuts (toTxOuts finalUTxO) === sortTxOuts (toList result)
           _ -> pure False
       _ -> pure True
 
@@ -745,7 +745,7 @@ performClose party = do
     HeadIsClosed{} -> Just ()
     _ -> Nothing
 
-performFanout :: (MonadThrow m, MonadAsync m, MonadDelay m) => Party -> RunMonad m UTxO
+performFanout :: (MonadThrow m, MonadAsync m, MonadDelay m) => Party -> RunMonad m (Set (TxOut CtxUTxO))
 performFanout party = do
   nodes <- gets nodes
   let thisNode = nodes ! party
@@ -753,13 +753,13 @@ performFanout party = do
   party `sendsInput` Input.Fanout
   findInOutput thisNode (100 :: Int)
  where
-  findInOutput :: (MonadDelay m, MonadThrow m) => TestHydraClient Tx m -> Int -> RunMonad m UTxO
+  findInOutput :: (MonadDelay m, MonadThrow m) => TestHydraClient Tx m -> Int -> RunMonad m (Set (TxOut CtxUTxO))
   findInOutput node n
     | n == 0 = failure "Failed to perform Fanout"
     | otherwise = do
         outputs <- lift $ serverOutputs node
         case find headIsFinalized outputs of
-          Just (HeadIsFinalized{utxo}) -> pure utxo
+          Just (HeadIsFinalized{finalizedUTxO}) -> pure finalizedUTxO
           _ -> lift (threadDelay 1) >> findInOutput node (n - 1)
 
   headIsFinalized :: ServerOutput Tx -> Bool