+++ hydra-tx/src/Hydra/Tx/Fanout.hs
@@ -4,6 +4,7 @@ import Hydra.Cardano.Api
 import Hydra.Prelude
 
 import Cardano.Api.UTxO qualified as UTxO
+import Data.Set qualified as Set
 import Hydra.Contract.Head qualified as Head
 import Hydra.Contract.HeadState qualified as Head
 import Hydra.Contract.MintAction (MintAction (..))
@@ -250,7 +251,7 @@ observeFanoutTx utxo tx = do
 
 data PartialFanoutObservation = PartialFanoutObservation
   { headId :: HeadId
-  , distributedUTxO :: UTxO
+  , distributedOutputs :: Set (TxOut CtxUTxO)
   }
   deriving stock (Eq, Show, Generic)
   deriving anyclass (ToJSON, FromJSON)
@@ -271,10 +272,8 @@ observePartialFanoutTx utxo tx = do
       Head.PartialFanout{numberOfPartialOutputs} -> do
         -- First output is the continuing head output, distributed outputs follow
         let numDistributed = fromIntegral numberOfPartialOutputs
-        let distributedOutputs = take numDistributed (drop 1 (txOuts' tx))
-        -- TxIn indices start at 0 for the continuing output, so distributed are at [1..n]
-        let distributedUTxO = UTxO.fromList $ zip (mkTxIn tx <$> [1 ..]) (toCtxUTxOTxOut <$> distributedOutputs)
-        pure PartialFanoutObservation{headId, distributedUTxO}
+        let distributedOutputs = Set.fromList (toCtxUTxOTxOut <$> take numDistributed (drop 1 (txOuts' tx)))
+        pure PartialFanoutObservation{headId, distributedOutputs}
       _ -> Nothing
 
 -- | Identify a final partial fanout tx by looking up the input spending the Head